requirements.txt
amqp==5.0.9
aniso8601==9.0.1
asgiref==3.6.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
boto3==1.20.26
botocore==1.23.26
celery==5.2.3
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cron-descriptor==1.2.35
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.2.0
django-clearcache==1.2.1
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-editorjs-fields==0.2.7
django-filter==23.1
django-graphql-jwt==0.3.4
django-import-export==3.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-storages==1.12.3
django-taggit==4.0.0
django-timezone-field==5.0
django-versatileimagefield==3.0
et-xmlfile==1.1.0
flower==1.0.0
graphene==3.2.2
graphene-django==3.0.0
graphene-django-pagination @ https://github.com/instruct-br/graphene-django-pagination/archive/master.zip#sha256=d727fe2df6632823830d92858f5f98ae91e4fb0ecb252188c92a334515b1bcf0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
humanize==3.13.1
jmespath==0.10.0
kombu==5.2.3
lxml==4.9.2
markdown2==2.4.8
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
phonenumbers==8.13.9
Pillow==9.5.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyparsing==3.0.6
pyquery==2.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2021.3
PyYAML==6.0
readtime==2.0.0
redis==4.1.0
s3transfer==0.5.0
sentry-sdk==1.5.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.2
tablib==3.4.0
text-unidecode==1.3
tornado==6.1
tzdata==2021.5
urllib3==1.26.7
validators==0.20.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
